healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: d220c4eca27d192e8ab0d451258b221b9a0bdef9
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 5ffff8a300e2fe4a2864ec7833cc6df9c801f879
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 98255fe58d66868d51e0ecd920323e276950afe4
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 98152f62cf5a25ab3976929b36ac50d9c4b54cee
Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>