Reboot sequence: Unmount active apexes before unmounting /data
Having mounted apexes with loop back devices backing files on /data
partition will prevent clean unmount of it. Unmounting them and tearing
down loop devices should minimize the risk of that.

Note that it won't fix the issue completely, as there are a few (~2-3)
processes that keep restarting even after SIGKILL is sent. Which means
that they can still hold references to apexes on /data partition. But
in practice probability of this is quite low.

Test: adb reboot
Test: put tzdata apex in /data/apex/active && adb reboot
Bug: 158152940
Change-Id: I4624567b3d0f304dba4c6e37b77abd89e57411de
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 2cc44cdd73a531c058e558b37b5e1b6a69c2f07e
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 42088f5a338ea9218b94008f4df9637fde79bc5c
Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>